requirements.txt
# Core dependencies
python-telegram-bot>=20.3
gspread>=5.10.0
pandas>=2.0.3
Flask>=2.3.3
gunicorn>=21.2.0
python-dotenv>=1.0.0
psutil>=5.9.5

# Additional dependencies
requests>=2.31.0
typing-extensions>=4.7.1

# Security and performance
cryptography>=41.0.3
urllib3>=2.0.4

# Retry mechanism
tenacity>=8.2.3
